requirements.txt
beautifulsoup4==4.9.3
pyTelegramBotAPI==3.8.3
python-dotenv==0.19.0
requests_html==0.10.0
schedule==1.1.0
selenium==3.141.0
